[feat] 장바구니 갯수 변경 구현 완료

+++ src/main/java/com/shop/mall/dto/CartResponseDto.java
@@ -25,7 +25,7 @@ public class CartResponseDto {
     @Getter
     @AllArgsConstructor
     public static class Ea {
-        private int cartId;
+        private Long cartId;
         private int totalprice;
     }
 }
